+++ pkg/engine/mutate/generatePatches.go
@@ -7,13 +7,15 @@ import (
 	"github.com/mattbaird/jsonpatch"
 )
 
-func generatePatches(src string, dst string) ([]jsonpatch.JsonPatchOperation, error) {
+func generatePatches(src, dst []byte) ([]jsonpatch.JsonPatchOperation, error) {
 	t := time.Now()
 	defer fmt.Printf("finished in %v\n", time.Since(t).String())
 
-	pp, err := jsonpatch.CreatePatch(baseBytes, expectBytes)
+	// pp, err := jsonpatch.CreatePatch(baseBytes, expectBytes)
+	pp, err := jsonpatch.CreatePatch(src, dst)
 	for _, p := range pp {
 		fmt.Println(p)
 	}
+
 	return pp, err
 }
